feat
백엔드에서 공통코드 업데이트 가능

+++ BE/src/main/java/FIXER/FIXER_BE/repository/BrandDeviceMapRepository.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface BrandDeviceMapRepository extends JpaRepository<BrandDeviceMap, Long> {
+public interface BrandDeviceMapRepository extends JpaRepository<BrandDeviceMap, Integer> {
 
     @Query("SELECT new FIXER.FIXER_BE.dto.CommonCodeDTO(dt.id, dt.name) " +
             "FROM DeviceType dt JOIN BrandDeviceMap bdm ON dt.id = bdm.deviceType.id " +
